org: new scripts folders
