+++ Assets/Scripts/CubeController.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0fba70e77b0055c43b7b290336bf980e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
